backend/pyproject.toml
[project]
name = "auth-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.0",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
